# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "passports",
        "title": "passports",
        "contact": {},
        "version": "0.0-0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/passports": {
            "get": {
                "description": "Возвращает все доступные паспорта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Паспорта"
                ],
                "summary": "Получение всех паспортов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название паспорта или его часть",
                        "name": "orbit_name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found",
                        "schema": {
                            "type": ""
                        }
                    }
                }
            }
        },
        "/passports/{passport_name}": {
            "get": {
                "description": "Возвращает подробную информацию о паспорте по его названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Паспорта"
                ],
                "summary": "Получение детализированной информации о паспорте",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название паспорта",
                        "name": "orbit_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/ping/{name}": {
            "get": {
                "description": "very friendly response",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tests"
                ],
                "summary": "Show hello text",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.pingResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.pingResp": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        }
    }
}